requirement.txt
streamlit==1.47.1
pytesseract==0.3.13
opencv-python==4.10.0.84
Pillow==11.0.0
googletrans==4.0.0-rc1
gTTS==2.5.4
numpy==1.26.4
pandas==2.3.1
bottleneck==1.3.7
numexpr==2.8.8

